refactor: rename search input component
